feat(admin): add admin app
Add admin app with basic configuration and setup. The app uses SolidJS
and TailwindCSS. It includes a README file with instructions on how to
install and run the app. The app has a basic header with a link to the
SolidJS GitHub page. The app is built using Vite and can be deployed to
any static host provider. The app has a .gitignore file that ignores
the dist folder.